Moved central C++ features to universal header file
