+++ dev/ast/group_by.h
@@ -4,7 +4,7 @@
 #include <type_traits>  //  std::true_type, std::false_type
 #include <utility>  //  std::forward, std::move
 
-#include "../cxx_polyfill.h"
+#include "../functional/cxx_polyfill.h"
 
 namespace sqlite_orm {
     namespace internal {
