+++ dev/conditions.h
@@ -6,8 +6,8 @@
 #include <tuple>  //  std::tuple, std::tuple_size
 #include <sstream>  //  std::stringstream
 
-#include "start_macros.h"
-#include "cxx_polyfill.h"
+#include "functional/cxx_universal.h"
+#include "functional/cxx_polyfill.h"
 #include "type_traits.h"
 #include "collate_argument.h"
 #include "constraints.h"
