+++ dev/member_traits/member_traits.h
@@ -2,8 +2,8 @@
 
 #include <type_traits>  //  std::enable_if, std::is_function, std::true_type, std::false_type
 
-#include "../start_macros.h"
-#include "../cxx_polyfill.h"
+#include "../functional/cxx_universal.h"
+#include "../functional/cxx_polyfill.h"
 
 namespace sqlite_orm {
     namespace internal {
