+++ dev/row_extractor_builder.h
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <cstddef>  //  std::nullptr_t
-
+#include "functional/cxx_universal.h"
 #include "row_extractor.h"
 #include "mapped_row_extractor.h"
 
@@ -10,7 +9,7 @@ namespace sqlite_orm {
     namespace internal {
 
         template<class T>
-        row_extractor<T> make_row_extractor(std::nullptr_t) {
+        row_extractor<T> make_row_extractor(nullptr_t) {
             return {};
         }
