+++ dev/storage.h
@@ -18,7 +18,7 @@
 #include <optional>  // std::optional
 #endif  // SQLITE_ORM_OPTIONAL_SUPPORTED
 
-#include "cxx_functional_polyfill.h"
+#include "functional/cxx_functional_polyfill.h"
 #include "type_traits.h"
 #include "tuple_helper/tuple_filter.h"
 #include "tuple_helper/tuple_helper.h"
