+++ dev/table_info.h
@@ -2,7 +2,7 @@
 
 #include <string>  //  std::string
 
-#include "start_macros.h"
+#include "functional/cxx_universal.h"
 
 namespace sqlite_orm {
