+++ dev/type_traits.h
@@ -2,8 +2,8 @@
 
 #include <type_traits>
 
-#include "start_macros.h"
-#include "cxx_polyfill.h"
+#include "functional/cxx_universal.h"
+#include "functional/cxx_polyfill.h"
 
 namespace sqlite_orm {
     // C++ generic traits used throughout the library
